resolvers dir
