+++ apps/blog/src/server/resolvers/index.ts
@@ -1,0 +1,1 @@
+export { resolvers } from './resolvers';
