+++ apps/blog/src/server/resolvers/scalars.ts
@@ -1,0 +1,7 @@
+import { GraphQLScalarType } from 'graphql';
+import { DateTimeResolver } from 'graphql-scalars';
+
+type GraphqlScalarTypeMap = { [key: string]: GraphQLScalarType };
+export const scalars: GraphqlScalarTypeMap = {
+  DateTime: DateTimeResolver,
+};
